backend/requirements.txt
# Python Dependencies for Tech News Crawler

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Web scraping
scrapy==2.11.0
selenium==4.15.2
playwright==1.40.0

# Translation
googletrans==4.0.0rc1
# azure-ai-translation-text==1.0.0  # Uncomment if using Azure
# openai==1.3.0  # Uncomment if using OpenAI

# Appwrite SDK
appwrite==4.1.0

# Utilities
pillow==10.1.0  # For image processing
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
